check integrity of inputs queries

+++ api/dataHandler/queries/channels/deleteChannel.js
@@ -1,16 +1,28 @@
 const { connect } = require("../../../helpers/connect")
+const { idChannelSchema, validator } = require("../../../helpers/jsonShemaValidator")
 
 const deleteChannel = async id => {
-    try {
-        await connect
-            .query(`DELETE FROM message WHERE channel_id=($1)`, [id]);
-        await connect
-            .query(`DELETE FROM channel WHERE id=($1)`, [id]);
-            
-    } catch (error) {
-        console.log(error)
-        throw (error);
+    const idValid = validator.validate(id, idChannelSchema);
+    if (idValid.errors.length > 0) {
+        throw (idValid.errors);
     }
+    await connect
+        .query(`DELETE FROM message WHERE channel_id=($1)`, [id])
+        .then(res => {
+            if (res.rowCount === 0) {
+                throw {
+                    status: 500,
+                };
+            } return res;
+        });
+    await connect.query(`DELETE FROM channel WHERE id=($1)`, [id])
+        .then(res => {
+            if (res.rowCount === 0) {
+                throw {
+                    status: 500,
+                };
+            } return res;
+        });
 };
 
 module.exports = { deleteChannel };
